fix: rebase conflicts, stack to deep and break tests

+++ src/market-making/branches/VaultRouterBranch.sol
@@ -35,7 +35,7 @@ contract VaultRouterBranch {
             Collateral.Data memory collateral
         )
     {
-        Vault.Data storage vault = Vault.load(vaultId);
+        Vault.Data storage vault = Vault.load(uint128(vaultId));
 
         totalDeposited = vault.totalDeposited;
         depositCap = vault.depositCap;
